Change Makefile to include program
The new Makefile variable `PROGRAM` indicates which program is to be
executed. It takes file naming into account, and considers that the
scripts that are to be executed are `script_$(PROGRAM)_server.py` and
`script_$(PROGRAM)_client.py`.

Bash script `run_rul.sh` was also modified to take into account the
scripts directory, and not the execution directory.

`New execution` logging by `turbofan.py` was also removed in this
commit.