Add optional validation procedure to epoch
The validation procedure comes after learning with the training batches,
and updating all client models. The client is then responsible for
communicating to the server whether it has any validation data or not.

What to do with the validation insights is still to be defined, but most
likely, it will be used to determine which models to save.

RUL Engine dataloaders have been added a new parameter `num_workers`,
which adds parallelism to the dataloader procedure. This increased the
training speed by 4 times with the current configurations.