Add centralized Turbofan CNN execution script
This commit allows training the single CNN model in a single process. A
validation procedure has also been added to the turbofan module.

Validation and test datasets now resort to the training minima and
maxima to perform their normalizatoin in the `_pre_processing`
procedure.

Add a new test procedure for the centralized neural network.