Add median test procedure for RUL prediction
Add a testing procedure that predicts the RUL based on the median of the
set of predictions for all samples of a flight. This was done to remove
outlier RUL predictions.

Change validation concern from distributed learning framework to server
execution script. This provides a more customizable approach to
reporting validation results, and what error metrics to present.