Save model when it outperforms persisted model
Add new struct containing the info required to persist a model to disk.
This information is required to determine whether a model is:
comparable; and if the model has improved w.r.t. its validation results
to the current persisted model.

To decouple the store, compare and equivalence methods, these are
expected to be implemented by each model in the script.

Runtime and model configuration files are now supplied as yaml files.
When a configuration file is not supplied to the model, a the default
configuration is used (in the src/models directory).

These additions were only implemented in the centralized execution
script for the turbofan dataset.