Add validation aggregation mvp
To aggregate all distributed models, federated averaging was used. This commit
adds a new aggregation method that validates each client's model, and based on
the validaiton results, keeps the best performing model.

From the pool of clients training, the validation procedure assigns to each, at
random, a model to be validated. Provided the validation results, the server
then assumes as global model the best performing model.