Add isolated training scenario
In this scenario, we consider a non-collaborative airline, which trains
its model solely with the data it has access to. In this case, it will
only have access to a single engine's data.

Additional changes were made to the centralized models to persist the
data consistently. A folder hierarchy is used in the results directory,
to represent different training scenarios.