Add faulty client scenario
This scenario is used to see how well each aggregation method learns
when a client has "noisy" data. I.e. if considered, the model the faulty
client provides would have a negative impact on the global model.

Rename the scripts responsible for running our proposed aggregation
methods based on the 2 decisions being made: Whether a full or random
validation procedure; Whether the best model is selected or a softmax is
performed to obtain the global model.

Add an execution script to run multiple scenarios in sequence.